Move P2PKHData to P2PKH class
Avoid existing app fail import
