+++ lib/src/payments/index.dart
@@ -24,18 +24,3 @@ class PaymentData {
     return 'PaymentData{address: $address, hash: $hash, output: $output, signature: $signature, pubkey: $pubkey, input: $input, witness: $witness}';
   }
 }
-
-// Backward compatibility
-@Deprecated('The "P2PKHData" class is deprecated. Use the "PaymentData" package instead.')
-class P2PKHData extends PaymentData {
-  P2PKHData({address, hash, output, pubkey, input, signature, witness}) :
-    super(
-        address: address,
-        hash: hash,
-        output: output,
-        pubkey: pubkey,
-        input: input,
-        signature: signature,
-        witness: witness
-    );
-}
