+++ lib/src/payments/p2pkh.dart
@@ -98,3 +98,18 @@ isValidOutput(Uint8List data) {
       data[23] == OPS['OP_EQUALVERIFY'] &&
       data[24] == OPS['OP_CHECKSIG'];
 }
+
+// Backward compatibility
+@Deprecated('The "P2PKHData" class is deprecated. Use the "PaymentData" package instead.')
+class P2PKHData extends PaymentData {
+  P2PKHData({address, hash, output, pubkey, input, signature, witness}) :
+    super(
+        address: address,
+        hash: hash,
+        output: output,
+        pubkey: pubkey,
+        input: input,
+        signature: signature,
+        witness: witness
+    );
+}
